server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
docxtpl==0.16.7
python-docx==1.0.1
python-dotenv==1.0.1
flask-cors==4.0.0